// services/course/src/shared/interfaces/multer-file.interface.ts
/**
 * Interface for Multer uploaded file
 * Compatible with Express.Multer.File but provides better type safety
 */
export interface MulterFile {
  /** Field name specified in the form */
  fieldname: string;
  /** Name of the file on the user's computer */
  originalname: string;
  /** Encoding type of the file */
  encoding: string;
  /** Mime type of the file */
  mimetype: string;
  /** Size of the file in bytes */
  size: number;
  /** The folder to which the file has been saved */
  destination: string;
  /** The name of the file within the destination */
  filename: string;
  /** Location of the uploaded file */
  path: string;
  /** A Buffer of the entire file */
  buffer: Buffer;
}